backend/python/requirements.txt
flask_restplus==0.12.1
networkx==2.3
Flask_Cors==3.0.8
requests==2.18.4
Werkzeug==0.15.4
setuptools==39.0.1
typing==3.7.4
Flask==1.1.1
base58==1.0.3
GitPython==2.1.12
protobuf==3.9.0
hypothesis==4.31.0
matplotlib==3.1.1
pytest==5.0.1
pywws==19.7.1
oef==0.4.0
fetchai-ledger-api==0.9.0a2
